@@ -1,18 +1,101 @@
-[tool.black]
-line-length = 100
-target-version = ['py36', 'py37', 'py38']
-include = '\.pyi?$'
-exclude = '''
-/(
-    \.eggs
-  | \.git
-  | \.hg
-  | \.mypy_cache
-  | \.tox
-  | \.venv
-  | _build
-  | buck-out
-  | build
-  | dist
-)/
-'''
+[build-system]
+requires = [
+    # SPEC: https://setuptools.pypa.io/en/latest/userguide/
+    "setuptools >= 61.0",
+    "wheel",
+]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+exclude = ["docs", "tests*"]
+
+[tool.setuptools.dynamic]
+version = { attr = "outplan.__version__" }
+
+
+[tool.wheel]
+universal = true
+
+
+[project]
+name = "outplan"
+description = "Support nested experiment/namespace base on Facebook Planout"
+readme = "README.md"
+license = { file = "LICENSE" }
+urls = { Repository = "https://github.com/xiachufang/outplan" }
+dynamic = ["version"]
+requires-python = ">= 3.6"
+authors = [{ name = "x1ah", email = "gaoxiaoqiang@xiachufang.com" }]
+classifiers = [
+    "Development Status :: 5 - Production/Stable",
+    "License :: OSI Approved :: MIT License",
+    "Natural Language :: Chinese (Simplified)",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.7",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+]
+
+dependencies = [
+    'planout==0.6.0',
+    'six',
+    'typing-extensions',
+]
+
+[project.optional-dependencies]
+test = ["pytest", "gevent"]
+
+
+[tool.mypy]
+ignore_missing_imports = true
+strict_optional = true
+show_error_context = true
+show_column_numbers = true
+
+
+[tool.ruff]
+# modified from:  https://docs.astral.sh/ruff/configuration/
+# extra settings: https://docs.astral.sh/ruff/settings/
+# extra rules:    https://docs.astral.sh/ruff/rules/
+line-length = 120
+indent-width = 4
+target-version = "py37"
+fix = true
+
+[tool.ruff.lint]
+select = [
+    # pyflakes / *
+    "F",
+    # pycodestyle / *
+    "E",
+    "W",
+    # isort / *
+    "I",
+    # pyupgrade / *
+    "UP",
+    # pylint / *
+    "PL",
+    # pandas-vet / *
+    "PD",
+    # numpy-specific / *
+    "NPY",
+    # flake8 / flake8-print
+    "T",
+    # ruff checking
+    "RUF",
+]
+ignore = []
+unfixable = []
+
+[tool.ruff.lint.pylint]
+max-args = 9
+max-branches = 15
+
+[tool.ruff.lint.pyupgrade]
+keep-runtime-typing = true
+
+[tool.ruff.format]
+quote-style = "preserve"
+indent-style = "space"
+skip-magic-trailing-comma = false
+line-ending = "auto"
