@@ -1,10 +1,8 @@
-# coding: utf-8
-
-
-class GroupResultType(object):
+class GroupResultType:
     """Type of group item result"""
-    group = 1   # return result directly
-    layer = 2   # namespace nested
+
+    group = 1  # return result directly
+    layer = 2  # namespace nested
 
 
 class UserTagFilterType:
