@@ -1,9 +1,8 @@
-# coding: utf-8
-
 from .client import ExperimentGroupClient
 from .exceptions import ExperimentBaseError
 from .experiment import ExperimentItem, GroupItem, NamespaceItem
 
+__version__ = "2.2.0"
 
 __all__ = (
     "ExperimentGroupClient",
